+++ packages/grid/src/features/standard-table/stories/StandardTableStoryHelper.ts
@@ -1,3 +1,4 @@
+import { ResultListBannerState } from "./../../../../../elements/src/components/ui/banners/result-list-banner/ResultListBanner";
 import { addDays, format } from "date-fns";
 import { useCallback, useState } from "react";
 import { createColumnConfig } from "../config/StandardTableColumnConfig";